alpha/requirements.txt
# Qlib 量化平台
pyqlib>=0.9.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 机器学习
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.5.0

# 工具库
tqdm>=4.62.0
pyyaml>=5.4.0
loguru>=0.5.0
joblib>=1.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# 数据库 (可选)
pymongo>=3.12.0
sqlalchemy>=1.4.0

# 其他工具
python-dotenv>=0.19.0
fire>=0.4.0
